project/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
pymongo>=4.5.0
razorpay>=1.4.1
beautifulsoup4>=4.12.0
requests>=2.31.0

urllib3>=2.0.4
